+++ packages/thumbprint-react/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
--   [Minor] Add `onMonthChange` and `modifiers` props to the `DatePicker` component.
+-   [Minor] Add `onMonthChange`, `daysThemeDotIndicator`, and `daysThemeStrikeout` props to the `DatePicker` component.
 
 ## 0.5.3 - 2019-04-09